# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend for time registration, projects, and glossary management.",
        "title": "tireg API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/time-entries/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a single time entry owned by the authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "time-entry"
                ],
                "summary": "Get a time entry",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Time entry id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_gerarc_tireg_internal_time-registry_infrastructure_in_rest_v1_dto.TimeEntryResponseDTO"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_gerarc_tireg_internal_shared_infrastructure_in_rest.ErrorResponseDTO"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_gerarc_tireg_internal_shared_infrastructure_in_rest.ErrorResponseDTO"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates a time entry owned by the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "time-entry"
                ],
                "summary": "Update a time entry",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Time entry id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Time entry payload",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/github_com_gerarc_tireg_internal_time-registry_infrastructure_in_rest_v1_dto.TimeEntryRequestDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_gerarc_tireg_internal_time-registry_infrastructure_in_rest_v1_dto.TimeEntryResponseDTO"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_gerarc_tireg_internal_shared_infrastructure_in_rest.ErrorResponseDTO"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_gerarc_tireg_internal_shared_infrastructure_in_rest.ErrorResponseDTO"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_gerarc_tireg_internal_shared_infrastructure_in_rest.ErrorResponseDTO"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes a time entry owned by the authenticated user",
                "tags": [
                    "time-entry"
                ],
                "summary": "Delete a time entry",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Time entry id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_gerarc_tireg_internal_shared_infrastructure_in_rest.ErrorResponseDTO"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_gerarc_tireg_internal_shared_infrastructure_in_rest.ErrorResponseDTO"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_gerarc_tireg_internal_shared_infrastructure_in_rest.ErrorResponseDTO": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "details": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "string"
                }
            }
        },
...
        "github_com_gerarc_tireg_internal_time-registry_infrastructure_in_rest_v1_dto.TimeEntryRequestDTO": {
            "type": "object",
            "properties": {
                "date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "end": {
                    "type": "string"
                },
                "hours": {
                    "type": "number"
                },
                "issue_key": {
                    "type": "string"
                },
                "jira_worklog_id": {
                    "type": "string"
                },
                "project_label": {
                    "type": "string"
                },
                "start": {
                    "type": "string"
                },
                "type_key": {
                    "type": "string"
                }
            }
        },
        "github_com_gerarc_tireg_internal_time-registry_infrastructure_in_rest_v1_dto.TimeEntryResponseDTO": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "end": {
                    "type": "string"
                },
                "hours": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "issue_key": {
                    "type": "string"
                },
                "jira_worklog_id": {
                    "type": "string"
                },
                "project_label": {
                    "type": "string"
                },
                "start": {
                    "type": "string"
                },
                "type_key": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and the JWT access token returned by /api/v1/auth/login.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}